requirements.txt
groq
python-telegram-bot
requests